Rename ActivityBroadcastReceiver to StringBroadcastReceiver
