--- app/src/main/java/com/andela/movit/listeners/ActivityCallback.java
@@ -1,8 +1,0 @@
-package com.andela.movit.listeners;
-
-
-import com.google.android.gms.location.DetectedActivity;
-
-public interface ActivityCallback {
-    void onActivityDetected(String activityName);
-}
