+++ app/src/main/java/com/andela/movit/listeners/IncomingStringCallback.java
@@ -1,5 +1,5 @@
 package com.andela.movit.listeners;
 
-public interface PlaceNameCallback {
+public interface IncomingStringCallback {
     void onPlaceNameFound(String placeName);
 }
