+++ app/src/main/java/com/andela/movit/receivers/StringBroadcastReceiver.java
@@ -1,23 +1,25 @@
-package com.andela.movit.activityrecognition;
+package com.andela.movit.receivers;
 
 import android.content.BroadcastReceiver;
 import android.content.Context;
 import android.content.Intent;
 
 import com.andela.movit.config.Constants;
-import com.andela.movit.listeners.ActivityCallback;
+import com.andela.movit.listeners.IncomingStringCallback;
 
-public class ActivityBroadcastReceiver extends BroadcastReceiver {
+public class StringBroadcastReceiver extends BroadcastReceiver {
 
-    public void setActivityCallback(ActivityCallback callback) {
+    private IncomingStringCallback callback;
+
+    private String actionName;
+
+    public void setIncomingStringCallback(IncomingStringCallback callback) {
         this.callback = callback;
     }
-
-    private ActivityCallback callback;
 
     @Override
     public void onReceive(Context context, Intent intent) {
         String activityName = intent.getStringExtra(Constants.ACTIVITY_NAME.getValue());
-        callback.onActivityDetected(activityName);
+        callback.onStringArrive(activityName);
     }
 }
